+++ usermod_idotmatrix.cpp
@@ -47,7 +47,7 @@
 #endif
 
 static constexpr const char* IDOTMATRIX_RELEASE = "0.9.0";
-static constexpr const char* IDOTMATRIX_BUILD = "0.9.0-rc.3";
+static constexpr const char* IDOTMATRIX_BUILD = "0.9.0-rc.5";
 static constexpr uint8_t IDOTMATRIX_APP_RELEASE_MAJOR = 0x00;
 static constexpr uint8_t IDOTMATRIX_APP_RELEASE_MINOR = 0x09;